Rename fieldFile to fieldUpload.
